requirements.txt
gradio
diffusers>=0.27.0
diffuser
torch
numpy
fastapi
uvicorn
pydantic
transformers==4.38.1
accelerate
tiktoken
einops
transformers_stream_generator>=0.0.4
scipy
torchvision
tensorboard
matplotlib
pydub
zhipuai
safetensors
pillow
safetensors